requirements.txt
streamlit==1.29.0torch==2.1.0torchvision==0.16.0Pillow==10.1.0gtts==2.4.0